# apps/server/docs/swagger-spec.yaml
openapi: 3.0.0
info:
  title: Shopify Partner CRM API
  description: API for managing Shopify partner CRM operations including partner accounts, organizations, leads, and activities.
  version: '1.0.0'
  contact:
    name: API Support
    url: https://api.cartrabbit.com/support
    email: support@cartrabbit.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms
...
paths:
  /user:
    post:
      summary: Create a user
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: 'rajesh kumar'
                email:
                  type: string
                  example: 'example@gmail.com'
                authenticationMethod:
                  type: string
                  enum:
                    - MAGIC_LINK
                    - GOOGLE
                avatarUrl:
                  type: string
                  example: 'http://amazons3.com/avatar.jpg'
              required:
                - name
                - email
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'User created successfully.'
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 'ck2j49r2a0d1m0741gzhjo41v'
                      name:
                        type: string
                        example: 'John'
                      authenticationMethod:
                        type: string
                        enum:
                          - GOOGLE
                          - MAGIC_LINK
                      createdAt:
                        type: integer
                        format: int64
                        example: 872748344
                      deletedAt:
                        type: integer
                        format: int64
                        example: 124234123
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Invalid data for user. Please check the provided data.'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Unauthorized access. Please provide a valid API key.'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'An unexpected error occurred. Please try again later.'
  /user/{id}:
    delete:
      summary: Delete a user by ID
      tags:
        - User
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 'ck2j49r2a0d1m0741gzhjo41v'
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'User deleted successfully.'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Invalid user ID. Please check the provided ID.'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Unauthorized access. Please provide a valid API key.'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'User not found.'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'An unexpected error occurred while deleting the user.'
    patch:
      summary: Update a user's name by ID
      tags:
        - User
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 'ck2j49r2a0d1m0741gzhjo41v'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: 'Updated Name'
              required:
                - name
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'User name updated successfully.'
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 'ck2j49r2a0d1m0741gzhjo41v'
                      name:
                        type: string
                        example: 'Updated Name'
                      createdAt:
                        type: integer
                        format: int64
                        example: 872748344
                      deletedAt:
                        type: integer
                        format: int64
                        example: 124234123
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Invalid data for user update. Please check the provided data.'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Unauthorized access. Please provide a valid API key.'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'User not found.'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'An unexpected error occurred while updating the user.'
    get:
      summary: Retrieve a user by ID
      tags:
        - User
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: 'ck2j49r2a0d1m0741gzhjo41v'
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'User retrieved successfully.'
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 'ck2j49r2a0d1m0741gzhjo41v'
                      name:
                        type: string
                        example: 'John'
                      email:
                        type: string
                        example: 'john.doe@example.com'
                      authenticationMethod:
                        type: string
                        enum:
                          - MAGIC_LINK
                          - GOOGLE
                        example: 'GOOGLE'
                      avatarUrl:
                        type: string
                        example: 'http://amazons3.com/avatar.jpg'
                      createdAt:
                        type: integer
                        format: int64
                        example: 872748344
                      deletedAt:
                        type: integer
                        format: int64
                        example: 124234123
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Invalid user ID. Please check the provided ID.'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Unauthorized access. Please provide a valid API key.'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'User not found.'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'An unexpected error occurred while retrieving the user.'
...
  /template/user/{userId}:
    get:
      summary: Get all templates by user ID
      tags:
        - Template
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            example: 'user12345'
      responses:
        '200':
          description: Templates retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Templates retrieved successfully.'
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: 'tmpl12345'
                        html:
                          type: string
                          example: '<html><body><h1>Hello World</h1></body></html>'
                        createdAt:
                          type: integer
                          format: int64
                          example: 872748344
                        updatedAt:
                          type: integer
                          format: int64
                          example: 872748344
                        deletedAt:
                          type: integer
                          format: int64
                          example: 0
                        userId:
                          type: string
                          example: 'user12345'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Invalid user ID. Please check the provided ID.'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Unauthorized access. Please provide a valid API key.'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'No templates found for the given user.'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'An unexpected error occurred while retrieving templates.'
   # Leads
...